Move GenomicAssemblyParser from parsers module into assembly
