+++ src/main/java/module-info.java
@@ -2,6 +2,5 @@ module org.monarchinitiative.svart {
     exports org.monarchinitiative.svart;
     exports org.monarchinitiative.svart.assembly;
     exports org.monarchinitiative.svart.interval;
-    exports org.monarchinitiative.svart.parsers;
     exports org.monarchinitiative.svart.util;
 }
