+++ src/test/java/org/monarchinitiative/svart/impl/DefaultGenomicAssemblyTest.java
@@ -7,7 +7,7 @@ import org.monarchinitiative.svart.*;
 import org.monarchinitiative.svart.assembly.AssignedMoleculeType;
 import org.monarchinitiative.svart.assembly.GenomicAssembly;
 import org.monarchinitiative.svart.assembly.SequenceRole;
-import org.monarchinitiative.svart.parsers.GenomicAssemblyParser;
+import org.monarchinitiative.svart.assembly.GenomicAssemblyParser;
 
 import java.nio.file.Path;
 import java.util.List;
